Move mapEditor types to /redux
- The types do not belong to gql. They are rather types used in redux
