+++ ui/src/hooks/routes/useEditRouteGeometry.ts
@@ -6,11 +6,9 @@ import {
   UpdateRouteGeometryMutationVariables,
   useUpdateRouteGeometryMutation,
 } from '../../generated/graphql';
-import {
-  mapInfraLinksAlongRouteToGraphQL,
-  mapRouteStopsToStopSequence,
-  RouteGeometry,
-} from '../../graphql';
+import { mapInfraLinksAlongRouteToGraphQL } from '../../graphql';
+import { buildStopSequenceFromRouteStops } from '../../redux';
+import { RouteGeometry } from '../../redux/types';
 import { removeFromApolloCache, showDangerToastWithError } from '../../utils';
 import { useValidateRoute } from './useValidateRoute';
 
@@ -48,7 +46,7 @@ export const useEditRouteGeometry = () => {
       newJourneyPattern: {
         on_route_id: routeId,
         scheduled_stop_point_in_journey_patterns:
-          mapRouteStopsToStopSequence(routeStops),
+          buildStopSequenceFromRouteStops(routeStops),
       },
     };
