+++ ui/src/redux/mappers/index.ts
@@ -1,0 +1,2 @@
+export * from './mapEditor';
+export * from './storeType';
