+++ ui/src/redux/types/index.ts
@@ -1,0 +1,2 @@
+export * from './mapEditor';
+export * from './modalMap';
